# Optimization: convert vector slice indices to floats