# Description
<!-- What does this pull request (PR) do? Why is it necessary? -->
<!-- Tell us about your new feature, improvement, or fix! -->
<!-- If your change includes user interface changes, please add an image, or an animation "An image is worth a thousand words!" -->
<!-- You can use https://www.cockos.com/licecap/ or similar to create animations -->
Numpy array slicing with object dtypes is much slower than defined (float) slicing. Defining the dtype here gives us a major performance boost. 

This bit of code is nearly identical to `Points` where this performance improvement was applied in [this PR](https://github.com/napari/napari/pull/4648)


## Type of change
performance optimization

# References
<!-- What resources, documentation, and guides were used in the creation of this PR? -->
<!-- If this is a bug-fix or otherwise resolves an issue, reference it here with "closes #(issue)" -->
With the 2D benchmark, we don't see significant improvements (though it is no worse, either), but with the 3D benchmark, there are huge improvements, e.g. `175±4ms` vs `3.46±0.03ms`. 


Results on `main`:
```
(napari_bench) benchmark$ asv run --python=same --bench Vectors2DSuite.time_set_view_slice
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_opt_miniconda3_envs_napari_bench_bin_python
[ 50.00%] ··· Running (benchmark_vectors_layer.Vectors2DSuite.time_set_view_slice--).
[100.00%] ··· benchmark_vectors_layer.Vectors2DSuite.time_set_view_slice                                                                                                                                                                                   ok
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     19.5±0.04μs 
                 64      20.0±0.1μs 
                256      21.3±0.2μs 
                1024     24.3±0.4μs 
                4096    32.4±0.09μs 
               16384     64.5±0.8μs 
               65536      190±1μs   
              ======== =============

(napari_bench) benchmark$ asv run --python=same --bench Vectors3DSuite.time_set_view_slice
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_opt_miniconda3_envs_napari_bench_bin_python
[ 50.00%] ··· Running (benchmark_vectors_layer.Vectors3DSuite.time_set_view_slice--).
[100.00%] ··· benchmark_vectors_layer.Vectors3DSuite.time_set_view_slice                                                                                                                                                                                   ok
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16       570±5μs   
                 64       702±6μs   
                256       1.22±0ms  
                1024    3.29±0.05ms 
                4096     11.6±0.2ms 
               16384     44.2±0.2ms 
               65536      175±4ms   
              ======== =============
```

Results on this branch:
```
(napari_bench) benchmark$ asv run --python=same --bench Vectors2DSuite.time_set_view_slice
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_opt_miniconda3_envs_napari_bench_bin_python
[ 50.00%] ··· Running (benchmark_vectors_layer.Vectors2DSuite.time_set_view_slice--).
[100.00%] ··· benchmark_vectors_layer.Vectors2DSuite.time_set_view_slice                                                                                                                                                                                   ok
[100.00%] ··· ======== ============
               param1              
              -------- ------------
                 16     19.4±0.1μs 
                 64     20.3±0.2μs 
                256     20.8±0.1μs 
                1024    23.5±0.1μs 
                4096    31.8±0.1μs 
               16384    63.6±0.2μs 
               65536    189±0.5μs  
              ======== ============

(napari_bench) benchmark$ asv run --python=same --bench Vectors3DSuite.time_set_view_slice
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_opt_miniconda3_envs_napari_bench_bin_python
[ 50.00%] ··· Running (benchmark_vectors_layer.Vectors3DSuite.time_set_view_slice--).
[100.00%] ··· benchmark_vectors_layer.Vectors3DSuite.time_set_view_slice                                                                                                                                                                                   ok
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16       517±6μs   
                 64       532±6μs   
                256       533±3μs   
                1024      573±8μs   
                4096      713±7μs   
               16384    1.30±0.03ms 
               65536    3.46±0.03ms 
              ======== =============
```

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [x] all tests pass with my change

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] If I included new strings, I have used `trans.` to make them localizable.
      For more information see our [translations guide](https://napari.org/developers/translations.html).