@sofroniewn 2022-08-02
This is great @kcpevey - love seeing those benchmarks. This is independent of #4846 I think so we can merge.

I'm good to merge now, though maybe @kcpevey do you want to create a new issue with @andy-sweet suggestion here

> But I still have a refactor itch around making Points._slice_data and Vectors._slice_data share more code and comments when they're doing very similar things.

I also like this idea, but it's something that could come in a followup after this merges.

If that sounds good to you @kcpevey let me know and i'll merge today. thanks!!