@kcpevey 2022-08-02
@sofroniewn This can be merged whenever you're ready. I've moved @andy-sweet's comment into [an issue](https://github.com/napari/napari/issues/4896)